+++ src/StashCommands.ts
@@ -3,6 +3,7 @@
 import * as vscode from 'vscode';
 import Config from './Config';
 import StashGit from './StashGit';
+import StashLabels from './StashLabels';
 import StashNode from './StashNode';
 
 enum StashType {
@@ -18,10 +19,12 @@ export class StashCommands {
     private config: Config;
     private channel: vscode.OutputChannel;
     private stashGit: StashGit;
+    private stashLabels: StashLabels;
 
-    constructor(config: Config, channel: vscode.OutputChannel) {
+    constructor(config: Config, channel: vscode.OutputChannel, stashLabels: StashLabels) {
         this.config = config;
         this.channel = channel;
+        this.stashLabels = stashLabels;
         this.stashGit = new StashGit();
     }
 
@@ -59,7 +62,7 @@ export class StashCommands {
 
         params.push(`stash@{${node.index}}`);
 
-        this.exec(params, 'Stash popped');
+        this.exec(params, 'Stash popped', node);
     }
 
     /**
@@ -74,7 +77,7 @@ export class StashCommands {
 
         params.push(`stash@{${node.index}}`);
 
-        this.exec(params, 'Stash applied');
+        this.exec(params, 'Stash applied', node);
     }
 
     /**
@@ -88,7 +91,7 @@ export class StashCommands {
             `stash@{${node.index}}`
         ];
 
-        this.exec(params, 'Stash branched');
+        this.exec(params, 'Stash branched', node);
     }
 
     /**
@@ -101,7 +104,7 @@ export class StashCommands {
             `stash@{${node.index}}`
         ];
 
-        this.exec(params, 'Stash dropped');
+        this.exec(params, 'Stash dropped', node);
     }
 
     /**
@@ -118,16 +121,29 @@ export class StashCommands {
      *
      * @param params         the array of command parameters
      * @param successMessage the string message to show on success
+     * @param node           the involved node
      */
-    private exec(params: string[], successMessage: string): void {
+    private exec(params: string[], successMessage: string, node?: StashNode): void {
         this.stashGit.exec(params)
             .then(
                 (result) => {
-                    this.showDetails('success', result, successMessage);
+                    let hasConflict = false;
+                    for (const line of result.split('\n')) {
+                        if (line.startsWith('CONFLICT (content): ')) {
+                            hasConflict = true;
+                            break;
+                        }
+                    }
+                    if (!hasConflict) {
+                        this.showDetails('success', result, successMessage, node);
+                    }
+                    else {
+                        this.showDetails('warning', result, `${successMessage} with conflicts`, node);
+                    }
                 },
                 (error) => {
                     const excerpt = error.substring(error.indexOf(':') + 1).trim();
-                    this.showDetails('error', error, excerpt);
+                    this.showDetails('error', error, excerpt, node);
                 }
             )
             .catch((error) => {
@@ -142,37 +158,38 @@ export class StashCommands {
      * @param message     the string result message
      * @param description the optional string alert description
      */
-    private showDetails(type: string, message: string, description?: string): void {
+    private showDetails(type: string, message: string, description?: string, node?: StashNode): void {
         message = message.trim();
-
-        const resume = description || message;
-        const actions = message.length > 0
-            ? [{ title: 'Show log' }]
-            : [];
 
         if (this.config.settings.log.autoclear) {
             this.channel.clear();
         }
 
         if (message.length > 0) {
-            this.channel.appendLine(`${message}\n`);
+            const currentTime = new Date();
+            this.channel.append(`> ${currentTime}`);
+            if (node) {
+                this.channel.append(`: ${this.stashLabels.getEntryName(node)}`);
+            }
+            this.channel.appendLine(`\n${message}\n`);
         }
 
+        const resume = description || message;
+        const actions = message.length > 0 ? [{ title: 'Show log' }] : [];
+        const callback = (value) => {
+            if (typeof value !== 'undefined') {
+                this.channel.show(true);
+            }
+        };
+
         if (type === 'success') {
-            vscode.window.showInformationMessage(resume, ...actions)
-                .then((value) => {
-                    if (typeof value !== 'undefined') {
-                        this.channel.show(true);
-                    }
-                });
+            vscode.window.showInformationMessage(resume, ...actions).then(callback);
+        }
+        else if (type === 'warning') {
+            vscode.window.showWarningMessage(resume, ...actions).then(callback);
         }
         else {
-            vscode.window.showErrorMessage(resume, ...actions)
-                .then((value) => {
-                    if (typeof value !== 'undefined') {
-                        this.channel.show(true);
-                    }
-                });
+            vscode.window.showErrorMessage(resume, ...actions).then(callback);
         }
     }
 }